chore: move non-oauth methods out of alby oauth service
