+++ alby/alby_oauth_service.go
@@ -51,9 +51,8 @@ const (
 )
 
 const (
-	albyOAuthAPIURL    = "https://api.getalby.com"
-	albyInternalAPIURL = "https://getalby.com/api"
-	albyOAuthAuthUrl   = "https://getalby.com/oauth"
+	albyOAuthAPIURL  = "https://api.getalby.com"
+	albyOAuthAuthUrl = "https://getalby.com/oauth"
 )
 
 const ALBY_ACCOUNT_APP_NAME = "getalby.com"
@@ -240,88 +239,6 @@ func (svc *albyOAuthService) fetchUserToken(ctx context.Context) (*oauth2.Token,
 
 	svc.saveToken(newToken)
 	return newToken, nil
-}
-
-func (svc *albyOAuthService) GetInfo(ctx context.Context) (*AlbyInfo, error) {
-	client := &http.Client{Timeout: 10 * time.Second}
-
-	req, err := http.NewRequest("GET", fmt.Sprintf("%s/internal/info", albyInternalAPIURL), nil)
-	if err != nil {
-		logger.Logger.WithError(err).Error("Error creating request to alby info endpoint")
-		return nil, err
-	}
-
-	setDefaultRequestHeaders(req)
-
-	res, err := client.Do(req)
-	if err != nil {
-		logger.Logger.WithError(err).Error("Failed to fetch /info")
-		return nil, err
-	}
-
-	type albyInfoHub struct {
-		LatestVersion      string `json:"latest_version"`
-		LatestReleaseNotes string `json:"latest_release_notes"`
-	}
-
-	type albyInfoIncident struct {
-		Name    string `json:"name"`
-		Started string `json:"started"`
-		Status  string `json:"status"`
-		Impact  string `json:"impact"`
-		Url     string `json:"url"`
-	}
-
-	type albyInfo struct {
-		Hub              albyInfoHub        `json:"hub"`
-		Status           string             `json:"status"`
-		Healthy          bool               `json:"healthy"`
-		AccountAvailable bool               `json:"account_available"` // false if country is blocked (can still use Alby Hub without an Alby Account)
-		Incidents        []albyInfoIncident `json:"incidents"`
-	}
-
-	body, err := io.ReadAll(res.Body)
-	if err != nil {
-		logger.Logger.WithError(err).Error("Failed to read response body")
-		return nil, errors.New("failed to read response body")
-	}
-
-	if res.StatusCode >= 300 {
-		logger.Logger.WithFields(logrus.Fields{
-			"body":        string(body),
-			"status_code": res.StatusCode,
-		}).Error("info endpoint returned non-success code")
-		return nil, fmt.Errorf("info endpoint returned non-success code: %s", string(body))
-	}
-
-	info := &albyInfo{}
-	err = json.Unmarshal(body, info)
-	if err != nil {
-		logger.Logger.WithError(err).Error("Failed to decode API response")
-		return nil, err
-	}
-
-	incidents := []AlbyInfoIncident{}
-	for _, incident := range info.Incidents {
-		incidents = append(incidents, AlbyInfoIncident{
-			Name:    incident.Name,
-			Started: incident.Started,
-			Status:  incident.Status,
-			Impact:  incident.Impact,
-			Url:     incident.Url,
-		})
-	}
-
-	return &AlbyInfo{
-		Hub: AlbyInfoHub{
-			LatestVersion:      info.Hub.LatestVersion,
-			LatestReleaseNotes: info.Hub.LatestReleaseNotes,
-		},
-		Status:           info.Status,
-		Healthy:          info.Healthy,
-		AccountAvailable: info.AccountAvailable,
-		Incidents:        incidents,
-	}, nil
 }
 
 func (svc *albyOAuthService) GetVssAuthToken(ctx context.Context, nodeIdentifier string) (string, error) {
@@ -1201,48 +1118,6 @@ func (svc *albyOAuthService) activateAlbyAccountNWCNode(ctx context.Context, wal
 	return nil
 }
 
-func (svc *albyOAuthService) GetChannelPeerSuggestions(ctx context.Context) ([]ChannelPeerSuggestion, error) {
-	client := &http.Client{Timeout: 10 * time.Second}
-
-	req, err := http.NewRequest("GET", fmt.Sprintf("%s/internal/channel_suggestions", albyInternalAPIURL), nil)
-	if err != nil {
-		logger.Logger.WithError(err).Error("Error creating request to channel_suggestions endpoint")
-		return nil, err
-	}
-
-	setDefaultRequestHeaders(req)
-
-	res, err := client.Do(req)
-	if err != nil {
-		logger.Logger.WithError(err).Error("Failed to fetch channel_suggestions endpoint")
-		return nil, err
-	}
-
-	body, err := io.ReadAll(res.Body)
-	if err != nil {
-		logger.Logger.WithError(err).Error("Failed to read response body")
-		return nil, errors.New("failed to read response body")
-	}
-
-	if res.StatusCode >= 300 {
-		logger.Logger.WithFields(logrus.Fields{
-			"body":        string(body),
-			"status_code": res.StatusCode,
-		}).Error("channel suggestions endpoint returned non-success code")
-		return nil, fmt.Errorf("channel suggestions endpoint returned non-success code: %s", string(body))
-	}
-
-	var suggestions []ChannelPeerSuggestion
-	err = json.Unmarshal(body, &suggestions)
-	if err != nil {
-		logger.Logger.WithError(err).Errorf("Failed to decode API response")
-		return nil, err
-	}
-
-	logger.Logger.WithFields(logrus.Fields{"channel_suggestions": suggestions}).Debug("Alby channel peer suggestions response")
-	return suggestions, nil
-}
-
 func (svc *albyOAuthService) GetLSPChannelOffer(ctx context.Context) (*LSPChannelOffer, error) {
 	token, err := svc.fetchUserToken(ctx)
 	if err != nil {
@@ -1289,64 +1164,6 @@ func (svc *albyOAuthService) GetLSPChannelOffer(ctx context.Context) (*LSPChanne
 	}
 
 	return lspChannelOffer, nil
-}
-
-func (svc *albyOAuthService) GetBitcoinRate(ctx context.Context) (*BitcoinRate, error) {
-	client := &http.Client{Timeout: 10 * time.Second}
-	currency := svc.cfg.GetCurrency()
-
-	url := fmt.Sprintf("%s/rates/%s", albyInternalAPIURL, currency)
-
-	req, err := http.NewRequest("GET", url, nil)
-	if err != nil {
-		logger.Logger.WithFields(logrus.Fields{
-			"currency": currency,
-			"error":    err,
-		}).Error("Error creating request to Bitcoin rate endpoint")
-		return nil, err
-	}
-	setDefaultRequestHeaders(req)
-
-	res, err := client.Do(req)
-	if err != nil {
-		logger.Logger.WithFields(logrus.Fields{
-			"currency": currency,
-			"error":    err,
-		}).Error("Failed to fetch Bitcoin rate from API")
-		return nil, err
-	}
-
-	defer res.Body.Close()
-
-	body, err := io.ReadAll(res.Body)
-	if err != nil {
-		logger.Logger.WithError(err).WithFields(logrus.Fields{
-			"url": url,
-		}).Error("Failed to read response body")
-		return nil, errors.New("failed to read response body")
-	}
-
-	if res.StatusCode >= 300 {
-		logger.Logger.WithFields(logrus.Fields{
-			"currency":    currency,
-			"body":        string(body),
-			"status_code": res.StatusCode,
-		}).Error("Bitcoin rate endpoint returned non-success code")
-		return nil, fmt.Errorf("bitcoin rate endpoint returned non-success code: %s", string(body))
-	}
-
-	var rate = &BitcoinRate{}
-	err = json.Unmarshal(body, rate)
-	if err != nil {
-		logger.Logger.WithFields(logrus.Fields{
-			"currency": currency,
-			"body":     string(body),
-			"error":    err,
-		}).Error("Failed to decode Bitcoin rate API response")
-		return nil, err
-	}
-
-	return rate, nil
 }
 
 func (svc *albyOAuthService) RequestAutoChannel(ctx context.Context, lnClient lnclient.LNClient, isPublic bool) (*AutoChannelResponse, error) {
