+++ wails/wails_app.go
@@ -28,7 +28,7 @@ type WailsApp struct {
 func NewApp(svc service.Service) *WailsApp {
 	return &WailsApp{
 		svc:     svc,
-		api:     api.NewAPI(svc, svc.GetDB(), svc.GetConfig(), svc.GetKeys(), svc.GetAlbyOAuthSvc(), svc.GetEventPublisher()),
+		api:     api.NewAPI(svc, svc.GetDB(), svc.GetConfig(), svc.GetKeys(), svc.GetAlbySvc(), svc.GetAlbyOAuthSvc(), svc.GetEventPublisher()),
 		db:      svc.GetDB(),
 		appsSvc: apps.NewAppsService(svc.GetDB(), svc.GetEventPublisher(), svc.GetKeys(), svc.GetConfig()),
 	}
